Fix src_Core/CPU/RISCV_MBox.bsv to work-around Vivado synth bug (detail follows)
When compiled with -D MULT_SYNTH, we produce Verilog multiplication
expressions which Vivado is supposed to synthesize into circuits
connected to DSPs. In RV32, the largest such exprs are 64bx64b->64b.
In RV, the largest such exprs are 128bx128b->128b. For the latter,
Vivado synthesis crashes. Wrote alternative functions that do
multiplication in pieces, restricting it to 64bx64b->64b expressions.