CPU: Allow debug module to interrupt and step over WFI
This involves two changes. Firstly, we allow debugger halt request
(either stop or step) to cause a WFI to resume in the same way that an
interrupt does, which puts the CPU in the CPU_RUNNING state and then
proceeds as normal to halt.

Secondly, for stepping to work properly, we need to perform
fa_step_check in rl_stage1_WFI so that rg_step_count is updated for
rl_WFI_resume to fire, rather than only being updated in rl_stage1_WFI
itself which is too late. The code for updating the various stages' full
states and calling fa_step_check then matches the code for
rl_stage1_xRET and rl_trap_fetch, and in fact, other than logging,
rl_WFI_resume and rl_trap_fetch are now identical.