MMU_Cache: Only update State_and_CTag_CSet once we know there is no error
If we get a read response error, we currently have already updated the
state and tag for the victim way, and thus attempting the request again
will hit in the cache and not raise an exception. For the D-Cache, we
don't expect to see read errors, as only the known-valid DRAM range is
cached, but for the I-Cache, everything is cached, so jumps to unmapped
addresses will hit this case. In particular, the reset trap vector
(0x1000) is unmapped, so trapping early in boot will trap to the reset
vector, take an instruction access fault trying to fill the cache line
but modify it to be clean and tagged with 0x1000, and then when it
proceeds to handle this second trap it will hit in the cache and start
executing whatever was in that cache line before it was evicted, with a
PC of 0x1000. Instead, we should only update the state and tag once we
have seen the first reponse for the burst and know that there is no
error.