CPU: Further fix single-step logic in the presence of branch mispredictions
Commit 6dd2445e added a redirection check to stage1_halted. However,
this condition is intending to check when the pipeline can be halted
after the stepped instruction has left stage 1, waiting for the next
instruction so that we can read the PC to retry at from stage 1. There
is no need to keep the pipeline running if redirecting and stage F is
busy; rl_stage1_stop already is guarded by stage F not being busy,
regardless of whether a redirection is in progress. Moreover, the
previous commit likely did not fix the bug it was intended to, since
although rl_pipe can run, halting will be true, so the instruction will
remain in stage 1 rather than progressing and then rl_stage1_stop will
eventually fire, just later than before.

Instead, we should move the logic to fa_step_check, and not update
rg_step_count to 1 for the instruction being stepped over until it can
redirect.