Provide an optional AXI4-Lite slave for coherent DMA
This introduces an AXI4-Lite slave interface at the top level that feeds
its requests through the same L1 D-cache as the CPU, which will be used
to provide coherent DMA in the Connectal-based AWS F1 SoC variant.

These incoming requests are fulfilled by putting them through a new
arbiter that sits between the core and the cache. This arbiter is only
present when the AXI4-Lite slave is enabled, but still adds no
additional latency to requests except under contention, with all ISA
tests taking the same number of cycles in simulation regardless of
whether the arbiter is present.